This is used for testing, so rename it to state so.
