+++ Tests/RuntimeTests/RuntimeTests.csproj
@@ -178,6 +178,7 @@
     <Compile Include="AssistTests\TestHelpers.cs" />
     <Compile Include="AssistTests\ServiceTests.cs" />
     <Compile Include="AssistTests\WorkingExampleOfValueRetrieverAndComparerAddition.cs" />
+    <Compile Include="Infrastructure\TestingBindingRuntime.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\Generator\TechTalk.SpecFlow.Generator.csproj">
